requirements.txt
Django==1.9.4
django-rest-swagger==0.3.5
djangorestframework==3.3.2
psycopg2==2.6.1
PyYAML==3.11
factory-boy==2.3.1
